requirements.txt
streamlit
openai
python-dotenv
extra-streamlit-components